Fixed timing problem for 2waycomcheck
Because the time since startup was stored as the number of milliseconds into an unsigned int, it would overflow at roughly 60 seconds, meaning that the main computer, ZEUS, had to run through the startup sequence in max. 1 minute from the boot of JAMES. Especially while testing in tethered mode, this was annoyingly short. This commit changes this timeframe for roughly 1 minute to roughly 100 min.